finish hook rotation sync

+++ giapnh/Assets/PQAssets/Scripts/Panel/OnlineGamePanel.cs
@@ -86,7 +86,7 @@ public class OnlineGamePanel : MonoBehaviour {
 			string rotation_str = cmd.getString(ArgCode.ARG_DROP_ROTATION,"");
 			Debug.Log(rotation_str);
 			string[] tab = rotation_str.Split(',');
-			Vector3 rotation = new Vector3(int.Parse(tab[0]),int.Parse(tab[1]),int.Parse(tab[2]));
+			Vector3 rotation = new Vector3(float.Parse(tab[0]),float.Parse(tab[1]),float.Parse(tab[2]));
 			Vector3 velocity = new Vector3(angel_x/100, angel_y/100, 0);
 			
 			Vector3 center_point = player.transform.position + new Vector3(0,-0.15f,0);
